infra/requirements.txt
aws_cdk.core==1.53.0
aws_cdk.aws_lambda==1.53.0
aws_cdk.aws_s3==1.53.0
path